+++ src/boot/ksdk1.1.0/devCCS811.c
@@ -239,8 +239,8 @@ readSensorRegisterCCS811(uint8_t deviceRegister, int numberOfBytes)
 void
 printSensorDataCCS811(bool hexModeFlag)
 {
-	uint8_t		readSensorRegisterValueLSB;
-	uint8_t		readSensorRegisterValueMSB;
+	uint16_t	readSensorRegisterValueLSB;
+	uint16_t	readSensorRegisterValueMSB;
 	int16_t		readSensorRegisterValueCombined;
 	int16_t		equivalentCO2, TVOC;
 	WarpStatus	i2cReadStatus;
