+++ src/boot/ksdk1.1.0/devMMA8451Q.c
@@ -199,8 +199,8 @@ readSensorRegisterMMA8451Q(uint8_t deviceRegister, int numberOfBytes)
 void
 printSensorDataMMA8451Q(bool hexModeFlag)
 {
-	uint8_t		readSensorRegisterValueLSB;
-	uint8_t		readSensorRegisterValueMSB;
+	uint16_t	readSensorRegisterValueLSB;
+	uint16_t	readSensorRegisterValueMSB;
 	int16_t		readSensorRegisterValueCombined;
 	WarpStatus	i2cReadStatus;
